Introduce stacked rank goals and migrate calories, trials, and MFC points Remove some no-longer-needed rank goals Implemented all goals except substitutions

+++ app/src/main/java/com/perrigogames/life4/android/AndroidPlatformStrings.kt
@@ -57,7 +57,7 @@ class AndroidPlatformStrings: PlatformStrings, KoinComponent {
             else c.getString(R.string.rank_goal_clear_trial, c.getString(rank.nameRes), count)
         }
 
-        override fun getMFCPointString(count: Int): String = c.getString(R.string.rank_goal_mfc_points, count)
+        override fun getMFCPointString(count: Double): String = c.getString(R.string.rank_goal_mfc_points, count)
 
         override fun folderString(
             folderCount: Int,
